Fact 96 - changing postcode search to a get request
